+++ service/common/api/urls.js
@@ -1,5 +1,0 @@
-/**
- * API_URL命名规则：API_模块_方法
- */
-// 登录
-export const API_RBAC_LOGIN = '/je/rbac/cloud/login/accountLogin';