+++ service/common/helper/je.js
@@ -1,31 +1,16 @@
-import { setupAxios } from './http';
 import * as Vue from 'vue';
 import * as Utils from '@jecloud/utils';
 import * as Ui from '@jecloud/ui';
 import { FuncUtil } from '@jecloud/func';
-import { logout, isLogin } from './system';
+import { CLI_ENVS } from '@common/helper/constant';
 /**
  * 使用全局工具类JE
  *
  * @export
  * @return { $vue,$i18n,$router,...utils }
  */
-const JE = {
-  useUi() {
-    return Ui;
-  },
-  useUtils() {
-    return Utils;
-  },
-  useVue() {
-    return Vue;
-  },
-  useSystem() {
-    return { logout, isLogin, ...FuncUtil, watchWebSocket() {} };
-  },
-};
 export function useJE() {
-  return JE;
+  return Utils.useJE();
 }
 
 /**
@@ -33,8 +18,9 @@ export function useJE() {
  * 页面可以通过JE.调用utils里的所有工具函数
  */
 export async function setupJE(vue) {
-  setupAxios();
-  mixinJE({ $vue: vue });
+  const JE = Utils.setupJE(vue, { Ui, Utils, Vue, Func: FuncUtil, CLI_ENVS });
+  // 绑定全局JE
+  window.JE = window.JE || JE;
 }
 
 /**
@@ -45,5 +31,5 @@ export async function setupJE(vue) {
  * @return {*}
  */
 export function mixinJE(object) {
-  return Object.assign(JE, object || {});
+  return Utils.mixinJE(object);
 }